refactor(55-04): barrel-preserving split of sdk-core folder/index.ts
- Extract fetchAndDecryptMetadata + loadFolderMetadata into load.ts
- Extract 4 pure transforms into metadata-ops.ts (rename/delete/addFilePointer/move)
- Extract IPNS record build + batch-publish fns into registration.ts
- Reduce index.ts to ~20 LoC barrel re-exporting all sibling modules
- registration.ts imports fetchAndDecryptMetadata directly from ./load (not via barrel)
- All 211 sdk-core tests pass; public export surface unchanged (D-05)
