+++ packages/sdk-core/src/folder/index.ts
@@ -6,615 +6,15 @@
  * Functions take metadata in, return updated metadata out.
  */
 
-import {
-  generateEd25519Keypair,
-  deriveIpnsName,
-  deriveEd25519PublicKey,
-  generateRandomBytes,
-  wrapKey,
-  bytesToHex,
-  hexToBytes,
-} from '@cipherbox/crypto';
-import {
-  encryptFolderMetadata,
-  decryptFolderMetadata,
-  createIpnsRecord,
-  marshalIpnsRecord,
-  type FolderMetadata,
-  type EncryptedFolderMetadata,
-  type FolderEntry,
-  type FolderChild,
-  type FilePointer,
-} from '@cipherbox/core';
-import type { SdkContext, TeeKeys } from '../types';
-import { addToIpfs, fetchFromIpfs } from '../ipfs';
-import { batchPublishIpnsRecords, resolveIpnsRecord } from '../ipns';
-import { withPerf } from '../perf';
-import type { FileIpnsRecordPayload } from '../file';
-import { publishWithCas } from '../cas';
-
 // Tree traversal utilities
 export { getDepth, calculateSubtreeDepth, isDescendantOf, type TreeNode } from './tree';
-import { mergeChildren } from './merge';
-export { mergeChildren };
+export { mergeChildren } from './merge';
 
-/**
- * Fetch and decrypt folder metadata from IPFS.
- *
- * @param cid - IPFS CID of the encrypted metadata blob
- * @param folderKey - Decrypted AES-256 folder key
- * @param ctx - SDK context for IPFS access
- * @returns Decrypted folder metadata (v2)
- */
-export async function fetchAndDecryptMetadata(
-  cid: string,
-  folderKey: Uint8Array,
-  ctx: SdkContext
-): Promise<FolderMetadata> {
-  return withPerf('folder:fetch-decrypt', async () => {
-    const encryptedBytes = await fetchFromIpfs(ctx, cid);
+// Load operations (fetch + decrypt from IPFS/IPNS)
+export * from './load';
 
-    // All folder metadata (including root) is v1 JSON {iv, data}.
-    // v2 blob format is only for the vault key blob (separate IPNS name).
-    const encryptedJson = new TextDecoder().decode(encryptedBytes);
-    const encrypted: EncryptedFolderMetadata = JSON.parse(encryptedJson);
-    return decryptFolderMetadata(encrypted, folderKey);
-  });
-}
+// Pure metadata transforms (rename, delete, add, move — no network side effects)
+export * from './metadata-ops';
 
-/**
- * Load a folder's metadata from IPNS.
- *
- * Resolves the folder's IPNS name to get the current metadata CID,
- * fetches and decrypts the metadata.
- *
- * @returns Decrypted folder metadata, sequence number, and CID, or null if IPNS not found
- */
-export async function loadFolderMetadata(params: {
-  ipnsName: string;
-  folderKey: Uint8Array;
-  ctx: SdkContext;
-}): Promise<{
-  metadata: FolderMetadata;
-  sequenceNumber: bigint;
-  cid: string;
-} | null> {
-  return withPerf('folder:load', async () => {
-    const resolved = await resolveIpnsRecord(params.ipnsName, params.ctx);
-    if (!resolved) return null;
-
-    const metadata = await fetchAndDecryptMetadata(resolved.cid, params.folderKey, params.ctx);
-
-    return {
-      metadata,
-      sequenceNumber: resolved.sequenceNumber,
-      cid: resolved.cid,
-    };
-  });
-}
-
-/**
- * Create a new subfolder with generated keys.
- *
- * Generates new Ed25519 IPNS keypair and AES-256 folder key,
- * wraps them with the user's public key for storage.
- *
- * @returns Created folder entry and decrypted keys
- */
-export async function createSubfolder(params: {
-  name: string;
-  userPublicKey: Uint8Array;
-  teeKeys?: TeeKeys;
-}): Promise<{
-  folder: FolderEntry;
-  ipnsPrivateKey: Uint8Array;
-  folderKey: Uint8Array;
-  encryptedIpnsPrivateKey?: string;
-  keyEpoch?: number;
-}> {
-  // 1. Generate Ed25519 keypair for folder IPNS
-  const ipnsKeypair = await generateEd25519Keypair();
-  const ipnsName = await deriveIpnsName(ipnsKeypair.publicKey);
-
-  // 2. Generate random AES-256 folder key
-  const folderKey = generateRandomBytes(32);
-
-  // 3. Wrap keys with user's public key (ECIES encryption)
-  const ipnsPrivateKeyEncrypted = bytesToHex(
-    await wrapKey(ipnsKeypair.privateKey, params.userPublicKey)
-  );
-  const folderKeyEncrypted = bytesToHex(await wrapKey(folderKey, params.userPublicKey));
-
-  // 4. TEE-02: Encrypt IPNS private key with TEE public key for republishing.
-  //    Wrap remaining steps in try/catch to zero key material on error.
-  //    On success the caller receives the keys, so we only zero on failure.
-  try {
-    let encryptedIpnsPrivateKey: string | undefined;
-    let keyEpoch: number | undefined;
-
-    if (params.teeKeys?.currentPublicKey) {
-      const teePublicKey = hexToBytes(params.teeKeys.currentPublicKey);
-      const encryptedKey = await wrapKey(ipnsKeypair.privateKey, teePublicKey);
-      encryptedIpnsPrivateKey = bytesToHex(encryptedKey);
-      keyEpoch = params.teeKeys.currentEpoch;
-    }
-
-    // 5. Create folder entry for parent's metadata
-    const now = Date.now();
-    const folder: FolderEntry = {
-      type: 'folder',
-      id: crypto.randomUUID(),
-      name: params.name,
-      ipnsName,
-      ipnsPrivateKeyEncrypted,
-      folderKeyEncrypted,
-      createdAt: now,
-      modifiedAt: now,
-    };
-
-    return {
-      folder,
-      ipnsPrivateKey: ipnsKeypair.privateKey,
-      folderKey,
-      encryptedIpnsPrivateKey,
-      keyEpoch,
-    };
-  } catch (error) {
-    ipnsKeypair.privateKey.fill(0);
-    folderKey.fill(0);
-    throw error;
-  }
-}
-
-/**
- * Update folder metadata and publish to IPNS.
- *
- * Encrypts the metadata with the folder key, uploads to IPFS,
- * and publishes an updated IPNS record pointing to the new CID.
- *
- * @returns New CID and sequence number
- */
-// T-47-01 / D-05 zeroization decision for updateFolderMetadataAndPublish:
-//
-// CALLER RETAINS OWNERSHIP — do NOT zero params.ipnsPrivateKey or params.folderKey here.
-//
-// Rationale (client.ts audit, A2): every call site passes live session keys from
-// the `folderTree` state (folder.ipnsKeypair.privateKey, folder.folderKey). These
-// keys are reused across the full session lifetime — the same folder may receive
-// many sequential publish calls (renameItem, deleteItem, uploadFile, moveItem, etc.).
-// Zeroing here would corrupt the in-memory session state and break all subsequent
-// operations on the same folder.
-//
-// Contrast with updateFileMetadata (sdk-core/file/index.ts:369-373), which DOES
-// zero because per-file IPNS keys are unwrapped fresh for each use and not stored
-// in long-lived state — that function IS the terminal consumer.
-//
-// The caller (sdk/src/client.ts) must own zeroing when it is safe to do so.
-// A guard test in __tests__/folder.test.ts asserts this documented non-zeroing
-// behavior so it cannot be accidentally changed.
-export async function updateFolderMetadataAndPublish(params: {
-  children: FolderChild[];
-  baseChildren?: FolderChild[];
-  folderKey: Uint8Array;
-  ipnsPrivateKey: Uint8Array;
-  ipnsPublicKey?: Uint8Array;
-  ipnsName: string;
-  sequenceNumber: bigint;
-  ctx: SdkContext;
-  encryptedIpnsPrivateKey?: string;
-  keyEpoch?: number;
-}): Promise<{ cid: string; newSequenceNumber: bigint; publishedChildren: FolderChild[] }> {
-  return withPerf('folder:update-publish', async () => {
-    // D-02: no base snapshot → warn once up front; the union fallback (mergeChildren
-    // with [] as base) still runs inside the merge callback below, but the warning is
-    // most useful when emitted regardless of whether a conflict occurs.
-    if (params.baseChildren === undefined) {
-      console.warn(
-        '[sdk-core] updateFolderMetadataAndPublish: baseChildren not provided for ' +
-          params.ipnsName +
-          ' — using union fallback (deletes may resurrect). Caller should pass baseChildren.'
-      );
-    }
-
-    // The base snapshot is owned here (REQ-3): the wrapper, not each caller, threads
-    // base → merge. Callers that omit baseChildren fall back to [] (union semantics).
-    const baseChildren = params.baseChildren ?? [];
-
-    const result = await publishWithCas<FolderChild[]>({
-      ipnsName: params.ipnsName,
-      ipnsPrivateKey: params.ipnsPrivateKey,
-      ipnsPublicKey: params.ipnsPublicKey,
-      sequenceNumber: params.sequenceNumber,
-      ctx: params.ctx,
-      encryptedIpnsPrivateKey: params.encryptedIpnsPrivateKey,
-      keyEpoch: params.keyEpoch,
-      maxAttempts: 4,
-      backoff: true,
-      // encrypt+upload happens per attempt so each gets a fresh CID (D-03)
-      encodeAndUpload: async (children) => {
-        const metadata: FolderMetadata = { version: 'v2', children };
-        const encrypted = await encryptFolderMetadata(metadata, params.folderKey);
-        const jsonStr = JSON.stringify(encrypted);
-        const encryptedBytes = new TextEncoder().encode(jsonStr);
-        const { cid } = await addToIpfs(params.ctx, encryptedBytes);
-        return cid;
-      },
-      decodeRemote: async (cid) => {
-        const remote = await fetchAndDecryptMetadata(cid, params.folderKey, params.ctx);
-        return remote.children;
-      },
-      // Three-way merge (D-01 / D-02). base is the snapshot owned by this wrapper.
-      merge: (base, local, remote) => ({
-        merged: mergeChildren(base ?? [], local, remote),
-      }),
-      localData: params.children,
-      baseData: baseChildren,
-    });
-
-    return {
-      cid: result.cid,
-      newSequenceNumber: result.newSequenceNumber,
-      publishedChildren: result.publishedData,
-    };
-  });
-}
-
-/**
- * Rename a child entry (folder or file) in folder metadata.
- *
- * Pure metadata operation: returns updated children array without publishing.
- */
-export function renameInFolder(params: {
-  children: FolderChild[];
-  childId: string;
-  newName: string;
-}): { updatedChildren: FolderChild[]; renamedChild: FolderChild } {
-  const children = [...params.children];
-  const index = children.findIndex((c) => c.id === params.childId);
-
-  if (index === -1) throw new Error('Item not found');
-
-  const nameExists = children.some((c) => c.name === params.newName && c.id !== params.childId);
-  if (nameExists) throw new Error('An item with this name already exists');
-
-  const renamedChild = {
-    ...children[index],
-    name: params.newName,
-    modifiedAt: Date.now(),
-  };
-  children[index] = renamedChild;
-
-  return { updatedChildren: children, renamedChild };
-}
-
-/**
- * Remove a child entry (folder or file) from folder metadata.
- *
- * Pure metadata operation: returns updated children array and the removed item.
- */
-export function deleteFromFolder(params: { children: FolderChild[]; childId: string }): {
-  updatedChildren: FolderChild[];
-  removedItem: FolderChild;
-} {
-  const index = params.children.findIndex((c) => c.id === params.childId);
-  if (index === -1) throw new Error('Item not found');
-
-  const removedItem = params.children[index];
-  const updatedChildren = params.children.filter((c) => c.id !== params.childId);
-
-  return { updatedChildren, removedItem };
-}
-
-/**
- * Add a file pointer to folder children.
- *
- * Pure metadata operation: returns updated children array with the new file pointer.
- */
-export function addFilePointerToFolder(params: {
-  children: FolderChild[];
-  fileId: string;
-  fileName: string;
-  fileMetaIpnsName: string;
-  ipnsPrivateKeyEncrypted: string;
-}): { updatedChildren: FolderChild[]; filePointer: FilePointer } {
-  const nameExists = params.children.some((c) => c.name === params.fileName);
-  if (nameExists) throw new Error('A file with this name already exists');
-
-  const now = Date.now();
-  const filePointer: FilePointer = {
-    type: 'file',
-    id: params.fileId,
-    name: params.fileName,
-    fileMetaIpnsName: params.fileMetaIpnsName,
-    ipnsPrivateKeyEncrypted: params.ipnsPrivateKeyEncrypted,
-    createdAt: now,
-    modifiedAt: now,
-  };
-
-  return {
-    updatedChildren: [...params.children, filePointer],
-    filePointer,
-  };
-}
-
-/**
- * Move a child entry between folders.
- *
- * Pure metadata operation: returns updated source and dest children arrays.
- * Uses add-before-remove pattern conceptually (caller publishes dest first, then source).
- */
-export function moveItem(params: {
-  sourceChildren: FolderChild[];
-  destChildren: FolderChild[];
-  childId: string;
-}): {
-  updatedSourceChildren: FolderChild[];
-  updatedDestChildren: FolderChild[];
-  movedItem: FolderChild;
-} {
-  const index = params.sourceChildren.findIndex((c) => c.id === params.childId);
-  if (index === -1) throw new Error('Item not found');
-
-  const movedItem = {
-    ...params.sourceChildren[index],
-    modifiedAt: Date.now(),
-  };
-
-  // Check name collision in destination
-  const nameExists = params.destChildren.some((c) => c.name === movedItem.name);
-  if (nameExists) {
-    throw new Error('An item with this name already exists in destination');
-  }
-
-  const updatedSourceChildren = params.sourceChildren.filter((c) => c.id !== params.childId);
-  const updatedDestChildren = [...params.destChildren, movedItem];
-
-  return { updatedSourceChildren, updatedDestChildren, movedItem };
-}
-
-// ---------------------------------------------------------------------------
-// File registration helpers (batch IPNS publish)
-// ---------------------------------------------------------------------------
-
-/** Safe base64 encoding that avoids call stack overflow from spread operator */
-function uint8ToBase64(bytes: Uint8Array): string {
-  let binary = '';
-  for (let i = 0; i < bytes.length; i++) {
-    binary += String.fromCharCode(bytes[i]);
-  }
-  return btoa(binary);
-}
-
-/**
- * Build a folder IPNS record payload for batch publishing.
- *
- * Creates and signs an IPNS record locally without publishing.
- * Returns the payload for inclusion in a batch publish call.
- */
-async function buildFolderIpnsRecord(params: {
-  children: FolderChild[];
-  folderKey: Uint8Array;
-  ipnsPrivateKey: Uint8Array;
-  ipnsPublicKey?: Uint8Array;
-  ipnsName: string;
-  sequenceNumber: bigint;
-  ctx: SdkContext;
-  encryptedIpnsPrivateKey?: string;
-  keyEpoch?: number;
-}): Promise<{
-  cid: string;
-  record: FileIpnsRecordPayload & { recordType: 'folder'; expectedSequenceNumber: string };
-  newSequenceNumber: bigint;
-}> {
-  const metadata: FolderMetadata = {
-    version: 'v2',
-    children: params.children,
-  };
-
-  const encrypted = await encryptFolderMetadata(metadata, params.folderKey);
-  const jsonStr = JSON.stringify(encrypted);
-  const encryptedBytes = new TextEncoder().encode(jsonStr);
-  const { cid } = await addToIpfs(params.ctx, encryptedBytes);
-
-  const newSeq = params.sequenceNumber + 1n;
-  const record = await createIpnsRecord(
-    params.ipnsPrivateKey,
-    `/ipfs/${cid}`,
-    newSeq,
-    24 * 60 * 60 * 1000
-  );
-  const recordBytes = marshalIpnsRecord(record);
-  const recordBase64 = uint8ToBase64(recordBytes);
-  const publicKey = uint8ToBase64(
-    params.ipnsPublicKey ?? deriveEd25519PublicKey(params.ipnsPrivateKey)
-  );
-
-  return {
-    cid,
-    record: {
-      ipnsName: params.ipnsName,
-      recordBase64,
-      publicKey,
-      metadataCid: cid,
-      encryptedIpnsPrivateKey: params.encryptedIpnsPrivateKey,
-      keyEpoch: params.keyEpoch,
-      recordType: 'folder',
-      expectedSequenceNumber: params.sequenceNumber.toString(),
-    },
-    newSequenceNumber: newSeq,
-  };
-}
-
-/**
- * Add a single file to a folder and batch-publish both IPNS records.
- *
- * Creates a FilePointer, builds a folder IPNS record, and publishes
- * both the file and folder IPNS records in a single batch API call.
- *
- * @returns Created file pointer and new folder sequence number
- * @throws Error if name collision exists or batch publish fails
- */
-export async function addFileToFolder(params: {
-  children: FolderChild[];
-  folderKey: Uint8Array;
-  ipnsPrivateKey: Uint8Array;
-  ipnsPublicKey?: Uint8Array;
-  ipnsName: string;
-  sequenceNumber: bigint;
-  fileId: string;
-  name: string;
-  fileIpnsRecord: FileIpnsRecordPayload;
-  ipnsPrivateKeyEncrypted: string;
-  ctx: SdkContext;
-}): Promise<{ filePointer: FilePointer; newSequenceNumber: bigint }> {
-  // 1. Check for name collision
-  const nameExists = params.children.some((c) => c.name === params.name);
-  if (nameExists) {
-    throw new Error('A file with this name already exists');
-  }
-
-  // 2. Create FilePointer
-  const now = Date.now();
-  const filePointer: FilePointer = {
-    type: 'file',
-    id: params.fileId,
-    name: params.name,
-    fileMetaIpnsName: params.fileIpnsRecord.ipnsName,
-    ipnsPrivateKeyEncrypted: params.ipnsPrivateKeyEncrypted,
-    createdAt: now,
-    modifiedAt: now,
-  };
-
-  // 3. Add FilePointer to parent's children
-  const children: FolderChild[] = [...params.children, filePointer];
-
-  // 4. Build folder IPNS record for batch publish
-  const folderResult = await buildFolderIpnsRecord({
-    children,
-    folderKey: params.folderKey,
-    ipnsPrivateKey: params.ipnsPrivateKey,
-    ipnsPublicKey: params.ipnsPublicKey,
-    ipnsName: params.ipnsName,
-    sequenceNumber: params.sequenceNumber,
-    ctx: params.ctx,
-  });
-
-  // 5. Batch publish: file IPNS record + folder IPNS record
-  const publishResult = await batchPublishIpnsRecords(
-    [{ ...params.fileIpnsRecord, recordType: 'file' as const }, folderResult.record],
-    params.ctx
-  );
-
-  if (publishResult.totalFailed > 0) {
-    throw new Error('Failed to publish one or more IPNS records');
-  }
-
-  return { filePointer, newSequenceNumber: folderResult.newSequenceNumber };
-}
-
-/**
- * Add multiple files to a folder and batch-publish all IPNS records.
- *
- * Creates FilePointers for each file, builds a single folder IPNS record,
- * and publishes all N file + 1 folder IPNS records in a single batch API call.
- *
- * @returns Created file pointers and new folder sequence number
- * @throws Error if any name collision exists or batch publish fails
- */
-export async function addFilesToFolder(params: {
-  children: FolderChild[];
-  folderKey: Uint8Array;
-  ipnsPrivateKey: Uint8Array;
-  ipnsPublicKey?: Uint8Array;
-  ipnsName: string;
-  sequenceNumber: bigint;
-  files: Array<{
-    fileId: string;
-    name: string;
-    fileIpnsRecord: FileIpnsRecordPayload;
-    ipnsPrivateKeyEncrypted: string;
-  }>;
-  ctx: SdkContext;
-}): Promise<{ filePointers: FilePointer[]; newSequenceNumber: bigint }> {
-  // 1. Build a set of existing child names for collision detection
-  const existingNames = new Set(params.children.map((c) => c.name));
-
-  // 2. Create FilePointer for each file, checking collisions
-  const filePointers: FilePointer[] = [];
-  const now = Date.now();
-
-  for (const file of params.files) {
-    if (existingNames.has(file.name)) {
-      throw new Error(`A file with name "${file.name}" already exists`);
-    }
-    existingNames.add(file.name);
-
-    const filePointer: FilePointer = {
-      type: 'file',
-      id: file.fileId,
-      name: file.name,
-      fileMetaIpnsName: file.fileIpnsRecord.ipnsName,
-      ipnsPrivateKeyEncrypted: file.ipnsPrivateKeyEncrypted,
-      createdAt: now,
-      modifiedAt: now,
-    };
-    filePointers.push(filePointer);
-  }
-
-  // 3. Build updated children array
-  const children: FolderChild[] = [...params.children, ...filePointers];
-
-  // 4. Build folder IPNS record
-  const folderResult = await buildFolderIpnsRecord({
-    children,
-    folderKey: params.folderKey,
-    ipnsPrivateKey: params.ipnsPrivateKey,
-    ipnsPublicKey: params.ipnsPublicKey,
-    ipnsName: params.ipnsName,
-    sequenceNumber: params.sequenceNumber,
-    ctx: params.ctx,
-  });
-
-  // 5. Batch publish: all N file IPNS records + 1 folder IPNS record
-  const allRecords = [
-    ...params.files.map((f) => ({ ...f.fileIpnsRecord, recordType: 'file' as const })),
-    folderResult.record,
-  ];
-  const publishResult = await batchPublishIpnsRecords(allRecords, params.ctx);
-
-  if (publishResult.totalFailed > 0) {
-    throw new Error('Failed to publish one or more IPNS records');
-  }
-
-  return { filePointers, newSequenceNumber: folderResult.newSequenceNumber };
-}
-
-/**
- * Replace file content in v2 metadata (content update).
- *
- * Publishes ONLY the file's IPNS record. The folder metadata is NOT
- * touched because the FilePointer still points to the same fileMetaIpnsName,
- * which now resolves to the updated content.
- *
- * @throws Error if file not found or publish fails
- */
-export async function replaceFileInFolder(params: {
-  children: FolderChild[];
-  fileId: string;
-  fileIpnsRecord: FileIpnsRecordPayload;
-  ctx: SdkContext;
-}): Promise<void> {
-  // 1. Verify file exists in parent's children
-  const fileExists = params.children.some((c) => c.type === 'file' && c.id === params.fileId);
-  if (!fileExists) throw new Error('File not found');
-
-  // 2. Publish ONLY the file IPNS record (folder metadata untouched!)
-  const publishResult = await batchPublishIpnsRecords(
-    [{ ...params.fileIpnsRecord, recordType: 'file' as const }],
-    params.ctx
-  );
-
-  if (publishResult.totalFailed > 0) {
-    throw new Error('Failed to publish file IPNS record');
-  }
-}
+// IPNS record build + batch-publish (createSubfolder, updateFolderMetadataAndPublish, etc.)
+export * from './registration';
